sat: zephyr: nrf: Refactor DTM code
Move the whole DTM code to a common file to avoid duplications.
