+++ port/zephyr/boards/nordic/nrf52/nrf52_soc.c
@@ -30,7 +30,7 @@ static nrfx_timer_t _timer0 = NRFX_TIMER_INSTANCE(NRF_TIMER_INST_GET(0));
 
 /* Keep track of power before and in sat tx mode */
 static nrf_radio_txpower_t _normal_power = RADIO_TXPOWER_TXPOWER_0dBm;
-static nrf_radio_txpower_t _sat_power = RADIO_TXPOWER_TXPOWER_0dBm;
+nrf_radio_txpower_t _sat_power = RADIO_TXPOWER_TXPOWER_0dBm;
 
 /**
  * Signature for APIs provided by the binary library.
@@ -196,88 +196,3 @@ int hubble_sat_soc_packet_send(const struct hubble_sat_packet_frames *packet)
 
 	return 0;
 }
-
-#ifdef CONFIG_HUBBLE_SAT_NETWORK_DTM_MODE
-
-int hubble_sat_soc_power_set(int8_t power)
-{
-	switch (power) {
-	case 8:
-		_sat_power = NRF_RADIO_TXPOWER_POS8DBM;
-		break;
-	case 7:
-		_sat_power = NRF_RADIO_TXPOWER_POS7DBM;
-		break;
-	case 6:
-		_sat_power = NRF_RADIO_TXPOWER_POS6DBM;
-		break;
-	case 5:
-		_sat_power = NRF_RADIO_TXPOWER_POS5DBM;
-		break;
-	case 4:
-		_sat_power = NRF_RADIO_TXPOWER_POS4DBM;
-		break;
-	case 3:
-		_sat_power = NRF_RADIO_TXPOWER_POS3DBM;
-		break;
-	case 2:
-		_sat_power = NRF_RADIO_TXPOWER_POS2DBM;
-		break;
-	case 0:
-		_sat_power = NRF_RADIO_TXPOWER_0DBM;
-		break;
-	case -4:
-		_sat_power = NRF_RADIO_TXPOWER_NEG4DBM;
-		break;
-	case -8:
-		_sat_power = NRF_RADIO_TXPOWER_NEG8DBM;
-		break;
-	case -12:
-		_sat_power = NRF_RADIO_TXPOWER_NEG12DBM;
-		break;
-	case -16:
-		_sat_power = NRF_RADIO_TXPOWER_NEG16DBM;
-		break;
-	case -20:
-		_sat_power = NRF_RADIO_TXPOWER_NEG20DBM;
-		break;
-	case -40:
-		_sat_power = NRF_RADIO_TXPOWER_NEG40DBM;
-		break;
-	default:
-		return -EINVAL;
-	}
-
-	nrf_radio_txpower_set(NRF_RADIO, _sat_power);
-	return 0;
-}
-
-int hubble_sat_soc_cw_start(uint8_t channel)
-{
-	int ret = hubble_nrf_lib_frequency_set(channel, 32);
-
-	if (ret != 0) {
-		return ret;
-	}
-
-	nrf_radio_task_trigger(NRF_RADIO, NRF_RADIO_TASK_TXEN);
-	while (!nrf_radio_event_check(NRF_RADIO, NRF_RADIO_EVENT_TXREADY)) {
-		/* Do nothing */
-	}
-
-	return 0;
-}
-
-int hubble_sat_soc_cw_stop(void)
-{
-	nrf_radio_task_trigger(NRF_RADIO, NRF_RADIO_TASK_DISABLE);
-
-	while (!nrf_radio_event_check(NRF_RADIO, NRF_RADIO_EVENT_DISABLED)) {
-		/* Do nothing */
-	}
-	nrf_radio_event_clear(NRF_RADIO, NRF_RADIO_EVENT_DISABLED);
-
-	return 0;
-}
-
-#endif /* CONFIG_HUBBLE_SAT_NETWORK_DTM_MODE */
